requirements.txt
appdirs==1.4.3
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
boto3==1.14.18
botocore==1.17.18
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
distlib==0.3.1
distro==1.4.0
Django==3.0.8
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
filelock==3.0.12
html5lib==1.0.1
identify==1.4.23
idna==2.8
install==1.3.3
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
nodeenv==1.4.0
packaging==20.3
pathspec==0.8.0
pep517==0.8.2
pre-commit==2.6.0
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.26
webencodings==0.5.1
